fix: 우선순위 오류 해결 : Priority type int -> double
- priority 자료형이 int여서 소수 계산 우선순위 비교가 되지 않는 문제 해결
- HRN 의 우선순위 계산이 정수끼리의 계산이여서 한 변수를 double로 캐스팅

+++ src/HRN.java
@@ -37,7 +37,7 @@ public class HRN extends CPUSchedulingAlgorithm {
             // priority = (대기 시간 + CPU 사용 시간) / CPU 사용시간
             for (Row row : availableRows) {
                 row.setWaitingTime(time - row.getArrivalTime());
-                row.setPriority((row.getWaitingTime() + row.getBurstTime()) / row.getBurstTime());
+                row.setPriority(((double)row.getWaitingTime() + row.getBurstTime()) / row.getBurstTime());
             }
 
             /*
